Update deploy script
This change includes the VMware-ovftool-4.2.0-5965791-lin.x86_64.bundle
file so that the OVFTool can be installed easily.

This change includes the final changes to the deploy script to push the
CoreOS channel image to the ESXi server. Everything has been tested on
ESXi and vSphere servers and proven to work in both scenarios.